# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MyMonastery API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Participant {
  name: string;
  email: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  createdAt: string;
}

export interface UpsertUserBody {
  name: string;
  email: string;
}

export type RitualFrequency =
  (typeof RitualFrequency)[keyof typeof RitualFrequency];

export const RitualFrequency = {
  weekly: "weekly",
  biweekly: "biweekly",
  monthly: "monthly",
} as const;

export type RitualStatus = (typeof RitualStatus)[keyof typeof RitualStatus];

export const RitualStatus = {
  on_track: "on_track",
  overdue: "overdue",
  needs_scheduling: "needs_scheduling",
} as const;

export interface Ritual {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  frequency: RitualFrequency;
  /** @nullable */
  dayPreference?: string | null;
  participants: Participant[];
  /** @nullable */
  intention?: string | null;
  ownerId: number;
  streak: number;
  /** @nullable */
  lastMeetupDate?: string | null;
  /** @nullable */
  nextMeetupDate?: string | null;
  status: RitualStatus;
  createdAt: string;
}

export type MeetupStatus = (typeof MeetupStatus)[keyof typeof MeetupStatus];

export const MeetupStatus = {
  planned: "planned",
  completed: "completed",
  skipped: "skipped",
} as const;

export interface Meetup {
  id: number;
  ritualId: number;
  scheduledDate: string;
  status: MeetupStatus;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export type MessageRole = (typeof MessageRole)[keyof typeof MessageRole];

export const MessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface Message {
  id: number;
  ritualId: number;
  role: MessageRole;
  content: string;
  createdAt: string;
}

export type RitualDetail = Ritual & {
  meetups?: Meetup[];
  messages?: Message[];
};

export type CreateRitualBodyFrequency =
  (typeof CreateRitualBodyFrequency)[keyof typeof CreateRitualBodyFrequency];

export const CreateRitualBodyFrequency = {
  weekly: "weekly",
  biweekly: "biweekly",
  monthly: "monthly",
} as const;

export interface CreateRitualBody {
  name: string;
  description?: string;
  frequency: CreateRitualBodyFrequency;
  dayPreference?: string;
  participants: Participant[];
  intention?: string;
  ownerId: number;
}

export type UpdateRitualBodyFrequency =
  (typeof UpdateRitualBodyFrequency)[keyof typeof UpdateRitualBodyFrequency];

export const UpdateRitualBodyFrequency = {
  weekly: "weekly",
  biweekly: "biweekly",
  monthly: "monthly",
} as const;

export interface UpdateRitualBody {
  name?: string;
  description?: string;
  frequency?: UpdateRitualBodyFrequency;
  dayPreference?: string;
  participants?: Participant[];
  intention?: string;
}

export type LogMeetupBodyStatus =
  (typeof LogMeetupBodyStatus)[keyof typeof LogMeetupBodyStatus];

export const LogMeetupBodyStatus = {
  completed: "completed",
  skipped: "skipped",
} as const;

export interface LogMeetupBody {
  status: LogMeetupBodyStatus;
  notes?: string;
  scheduledDate: string;
}

export interface SendMessageBody {
  content: string;
}