# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MyMonastery API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Get current user
 */
export const GetUserResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  email: zod.string(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Create or update user
 */
export const UpsertUserBody = zod.object({
  name: zod.string(),
  email: zod.string(),
});

export const UpsertUserResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  email: zod.string(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary List all rituals
 */
export const ListRitualsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string().nullish(),
  frequency: zod.enum(["weekly", "biweekly", "monthly"]),
  dayPreference: zod.string().nullish(),
  participants: zod.array(
    zod.object({
      name: zod.string(),
      email: zod.string(),
    }),
  ),
  intention: zod.string().nullish(),
  ownerId: zod.number(),
  streak: zod.number(),
  lastMeetupDate: zod.string().nullish(),
  nextMeetupDate: zod.string().nullish(),
  status: zod.enum(["on_track", "overdue", "needs_scheduling"]),
  createdAt: zod.coerce.date(),
});
export const ListRitualsResponse = zod.array(ListRitualsResponseItem);

/**
 * @summary Create a ritual
 */
export const CreateRitualBody = zod.object({
  name: zod.string(),
  description: zod.string().optional(),
  frequency: zod.enum(["weekly", "biweekly", "monthly"]),
  dayPreference: zod.string().optional(),
  participants: zod.array(
    zod.object({
      name: zod.string(),
      email: zod.string(),
    }),
  ),
  intention: zod.string().optional(),
  ownerId: zod.number(),
});

/**
 * @summary Get a ritual
 */
export const GetRitualParams = zod.object({
  id: zod.coerce.number(),
});

export const GetRitualResponse = zod
  .object({
    id: zod.number(),
    name: zod.string(),
    description: zod.string().nullish(),
    frequency: zod.enum(["weekly", "biweekly", "monthly"]),
    dayPreference: zod.string().nullish(),
    participants: zod.array(
      zod.object({
        name: zod.string(),
        email: zod.string(),
      }),
    ),
    intention: zod.string().nullish(),
    ownerId: zod.number(),
    streak: zod.number(),
    lastMeetupDate: zod.string().nullish(),
    nextMeetupDate: zod.string().nullish(),
    status: zod.enum(["on_track", "overdue", "needs_scheduling"]),
    createdAt: zod.coerce.date(),
  })
  .and(
    zod.object({
      meetups: zod
        .array(
          zod.object({
            id: zod.number(),
            ritualId: zod.number(),
            scheduledDate: zod.coerce.date(),
            status: zod.enum(["planned", "completed", "skipped"]),
            notes: zod.string().nullish(),
            createdAt: zod.coerce.date(),
          }),
        )
        .optional(),
      messages: zod
        .array(
          zod.object({
            id: zod.number(),
            ritualId: zod.number(),
            role: zod.enum(["user", "assistant"]),
            content: zod.string(),
            createdAt: zod.coerce.date(),
          }),
        )
        .optional(),
    }),
  );

/**
 * @summary Update a ritual
 */
export const UpdateRitualParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateRitualBody = zod.object({
  name: zod.string().optional(),
  description: zod.string().optional(),
  frequency: zod.enum(["weekly", "biweekly", "monthly"]).optional(),
  dayPreference: zod.string().optional(),
  participants: zod
    .array(
      zod.object({
        name: zod.string(),
        email: zod.string(),
      }),
    )
    .optional(),
  intention: zod.string().optional(),
});

export const UpdateRitualResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string().nullish(),
  frequency: zod.enum(["weekly", "biweekly", "monthly"]),
  dayPreference: zod.string().nullish(),
  participants: zod.array(
    zod.object({
      name: zod.string(),
      email: zod.string(),
    }),
  ),
  intention: zod.string().nullish(),
  ownerId: zod.number(),
  streak: zod.number(),
  lastMeetupDate: zod.string().nullish(),
  nextMeetupDate: zod.string().nullish(),
  status: zod.enum(["on_track", "overdue", "needs_scheduling"]),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Delete a ritual
 */
export const DeleteRitualParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get meetup history for a ritual
 */
export const ListMeetupsParams = zod.object({
  id: zod.coerce.number(),
});

export const ListMeetupsResponseItem = zod.object({
  id: zod.number(),
  ritualId: zod.number(),
  scheduledDate: zod.coerce.date(),
  status: zod.enum(["planned", "completed", "skipped"]),
  notes: zod.string().nullish(),
  createdAt: zod.coerce.date(),
});
export const ListMeetupsResponse = zod.array(ListMeetupsResponseItem);

/**
 * @summary Log a meetup
 */
export const LogMeetupParams = zod.object({
  id: zod.coerce.number(),
});

export const LogMeetupBody = zod.object({
  status: zod.enum(["completed", "skipped"]),
  notes: zod.string().optional(),
  scheduledDate: zod.coerce.date(),
});

/**
 * @summary Get chat messages for a ritual
 */
export const ListMessagesParams = zod.object({
  id: zod.coerce.number(),
});

export const ListMessagesResponseItem = zod.object({
  id: zod.number(),
  ritualId: zod.number(),
  role: zod.enum(["user", "assistant"]),
  content: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListMessagesResponse = zod.array(ListMessagesResponseItem);

/**
 * @summary Send a message to the AI coordinator
 */
export const SendMessageParams = zod.object({
  id: zod.coerce.number(),
});

export const SendMessageBody = zod.object({
  content: zod.string(),
});

export const SendMessageResponse = zod.object({
  id: zod.number(),
  ritualId: zod.number(),
  role: zod.enum(["user", "assistant"]),
  content: zod.string(),
  createdAt: zod.coerce.date(),
});