# lib/api-zod/src/generated/types/createRitualBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MyMonastery API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateRitualBodyFrequency } from "./createRitualBodyFrequency";
import type { Participant } from "./participant";

export interface CreateRitualBody {
  name: string;
  description?: string;
  frequency: CreateRitualBodyFrequency;
  dayPreference?: string;
  participants: Participant[];
  intention?: string;
  ownerId: number;
}