# lib/api-zod/src/generated/types/meetup.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MyMonastery API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MeetupStatus } from "./meetupStatus";

export interface Meetup {
  id: number;
  ritualId: number;
  scheduledDate: Date;
  status: MeetupStatus;
  /** @nullable */
  notes?: string | null;
  createdAt: Date;
}