# lib/api-zod/src/generated/types/message.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MyMonastery API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MessageRole } from "./messageRole";

export interface Message {
  id: number;
  ritualId: number;
  role: MessageRole;
  content: string;
  createdAt: Date;
}