# lib/api-zod/src/generated/types/ritual.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MyMonastery API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Participant } from "./participant";
import type { RitualFrequency } from "./ritualFrequency";
import type { RitualStatus } from "./ritualStatus";

export interface Ritual {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  frequency: RitualFrequency;
  /** @nullable */
  dayPreference?: string | null;
  participants: Participant[];
  /** @nullable */
  intention?: string | null;
  ownerId: number;
  streak: number;
  /** @nullable */
  lastMeetupDate?: string | null;
  /** @nullable */
  nextMeetupDate?: string | null;
  status: RitualStatus;
  createdAt: Date;
}