# lib/api-zod/src/generated/types/ritualDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MyMonastery API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Meetup } from "./meetup";
import type { Message } from "./message";
import type { Ritual } from "./ritual";

export type RitualDetail = Ritual & {
  meetups?: Meetup[];
  messages?: Message[];
};