# lib/api-zod/src/generated/types/upsertUserBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MyMonastery API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpsertUserBody {
  name: string;
  email: string;
}